Code cleanup and organization
Moved classes into different packages for easier navigation
