+++ src/main/java/com/alexslo/bank/Application.java
@@ -1,0 +1,7 @@
+package com.alexslo.bank;
+
+public class Application {
+    public static void main(String[] args) {
+
+    }
+}
