+++ src/main/java/com/alexslo/bank/controller/servlets/AuthServlet.java
@@ -1,8 +1,8 @@
-package controller.servlets;
+package com.alexslo.bank.controller.servlets;
 
-import model.dao.impl.UserDao;
-import model.entity.Transaction;
-import model.entity.UserRole;
+import com.alexslo.bank.service.impl.UserDao;
+import com.alexslo.bank.model.entity.Transaction;
+import com.alexslo.bank.model.entity.UserRole;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -34,7 +34,7 @@ public class AuthServlet extends HttpServlet {
         final String login = req.getParameter("login");
         final String password = req.getParameter("password");
 
-        @SuppressWarnings("unchecked") final AtomicReference<UserDao> dao = (AtomicReference<UserDao>) req.getServletContext().getAttribute("dao");
+        @SuppressWarnings("unchecked") final AtomicReference<UserDao> dao = (AtomicReference<UserDao>) req.getServletContext().getAttribute("com/alexslo/bank/dao");
         req.setAttribute("transactions", transactions.values());
         final HttpSession session = req.getSession();
