+++ src/main/java/com/alexslo/bank/model/entity/AccountType.java
@@ -1,4 +1,4 @@
-package model.entity;
+package com.alexslo.bank.model.entity;
 
 public enum AccountType {
     DEPOSIT,CREDIT
