+++ src/main/java/com/alexslo/bank/model/entity/UserRole.java
@@ -1,4 +1,4 @@
-package model.entity;
+package com.alexslo.bank.model.entity;
 
 public enum UserRole {
     GUEST,USER,ADMIN;
