+++ src/main/java/com/alexslo/bank/service/impl/AccountDao.java
@@ -1,9 +1,9 @@
-package model.dao.impl;
+package com.alexslo.bank.service.impl;
 
-import model.entity.Account;
-import model.entity.AccountType;
+import com.alexslo.bank.model.entity.Account;
+import com.alexslo.bank.model.entity.AccountType;
 
-public class AccountDao implements model.dao.AccountDao {
+public class AccountDao implements com.alexslo.bank.dao.AccountDao {
 
     @Override
     public void createAccount(Account account) {
