--- src/main/java/controller/ContextListener.java
@@ -1,20 +1,0 @@
-package controller;
-
-
-import javax.servlet.ServletContextEvent;
-import javax.servlet.ServletContextListener;
-import javax.servlet.annotation.WebListener;
-
-@WebListener
-public class ContextListener implements ServletContextListener {
-
-
-    @Override
-    public void contextInitialized(ServletContextEvent servletContextEvent) {
-
-    }
-
-    @Override
-    public void contextDestroyed(ServletContextEvent servletContextEvent) {
-    }
-}
