+++ src/main/webapp/WEB-INF/web.xml
@@ -7,7 +7,7 @@
     <!--filter check exist session for sign in-->
     <servlet>
         <servlet-name>AuthServlet</servlet-name>
-        <servlet-class>controller.servlets.AuthServlet</servlet-class>
+        <servlet-class>com.alexslo.bank.controller.servlets.AuthServlet</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>AuthServlet</servlet-name>
@@ -16,7 +16,7 @@
     <!--Logout servlet-->
     <servlet>
         <servlet-name>LogoutServlet</servlet-name>
-        <servlet-class>controller.servlets.LogoutServlet</servlet-class>
+        <servlet-class>com.alexslo.bank.controller.servlets.LogoutServlet</servlet-class>
     </servlet>
 
     <servlet-mapping>
@@ -26,7 +26,7 @@
 
     <servlet>
         <servlet-name>TransactionServlet</servlet-name>
-        <servlet-class>controller.servlets.TransactionServlet</servlet-class>
+        <servlet-class>com.alexslo.bank.controller.servlets.TransactionServlet</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>TransactionServlet</servlet-name>
